refactored forcepowers import
