+++ modules/importer/data-importer.js
@@ -188,7 +188,7 @@ export default class DataImporter extends FormApplication {
           promises.push(OggDude.Import.Weapons(xmlDoc, zip));
           promises.push(OggDude.Import.Armor(xmlDoc, zip));
           promises.push(OggDude.Import.Talents(xmlDoc, zip));
-          promises.push(this._handleForcePowers(xmlDoc, zip));
+          promises.push(OggDude.Import.ForcePowers(xmlDoc, zip));
           promises.push(OggDude.Import.SignatureAbilities(xmlDoc, zip));
           promises.push(OggDude.Import.ItemAttachments(xmlDoc));
         } else {
@@ -222,235 +222,6 @@ export default class DataImporter extends FormApplication {
       CONFIG.temporary = {};
       this.close();
     }
-  }
-
-  async _handleForcePowers(xmlDoc, zip) {
-    this._importLogger(`Starting Force Power Import`);
-    const forceabilities = xmlDoc.getElementsByTagName("ForceAbility");
-    if (forceabilities.length > 0) {
-      $(".import-progress.force").toggleClass("import-hidden");
-      let pack = await this._getCompendiumPack("Item", `oggdude.ForcePowers`);
-
-      const fa = JXON.xmlToJs(xmlDoc);
-      // now we need to loop through the files in the Force Powers folder
-
-      const forcePowersFiles = Object.values(zip.files).filter((file) => {
-        return !file.dir && file.name.split(".").pop() === "xml" && file.name.includes("/Force Powers/");
-      });
-
-      let totalCount = forcePowersFiles.length;
-      let currentCount = 0;
-      this._importLogger(`Beginning import of ${forcePowersFiles.length} force powers`);
-
-      await this.asyncForEach(forcePowersFiles, async (file) => {
-        try {
-          const data = await zip.file(file.name).async("text");
-          const xmlDoc1 = ImportHelpers.stringToXml(data);
-          const fp = JXON.xmlToJs(xmlDoc1);
-
-          // setup the base information
-
-          let power = {
-            name: fp.ForcePower.Name,
-            type: "forcepower",
-            flags: {
-              ffgimportid: fp.ForcePower.Key,
-            },
-            data: {
-              attributes: {},
-              upgrades: {},
-            },
-          };
-
-          this._importLogger(`Start importing force power ${fp.ForcePower.Name}`);
-
-          // get the basic power informatio
-          const importKey = fp.ForcePower.AbilityRows.AbilityRow[0].Abilities.Key[0];
-
-          let forceAbility = fa.ForceAbilities.ForceAbility.find((ability) => {
-            return ability.Key === importKey;
-          });
-
-          power.data.description = forceAbility.Description;
-          power.data.description += ImportHelpers.getSources(fp?.ForcePower?.Sources ?? fp?.ForcePower?.Source);
-
-          if (forceAbility?.DieModifiers?.DieModifier) {
-            if (!Array.isArray(forceAbility.DieModifiers.DieModifier)) {
-              forceAbility.DieModifiers.DieModifier = [forceAbility.DieModifiers.DieModifier];
-            }
-            forceAbility.DieModifiers.DieModifier.forEach((mod) => {
-              const attr = ImportHelpers.getBaseModAttributeObject({
-                Key: mod.SkillKey,
-                ...mod,
-              });
-              if (attr) {
-                power.data.attributes[attr.type] = attr.value;
-              }
-            });
-          }
-
-          // next we will parse the rows
-
-          for (let i = 1; i < fp.ForcePower.AbilityRows.AbilityRow.length; i += 1) {
-            try {
-              const row = fp.ForcePower.AbilityRows.AbilityRow[i];
-              row.Abilities.Key.forEach((keyName, index) => {
-                let rowAbility = {};
-
-                let rowAbilityData = fa.ForceAbilities.ForceAbility.find((a) => {
-                  return a.Key === keyName;
-                });
-
-                rowAbility.name = rowAbilityData.Name;
-                rowAbility.description = rowAbilityData.Description;
-                rowAbility.cost = row.Costs.Cost[index];
-                rowAbility.visible = true;
-                rowAbility.attributes = {};
-
-                if (row.Directions.Direction[index].Up) {
-                  rowAbility["links-top-1"] = true;
-                }
-
-                switch (row.AbilitySpan.Span[index]) {
-                  case "1":
-                    rowAbility.size = "single";
-                    break;
-                  case "2":
-                    rowAbility.size = "double";
-                    if (index < 3 && row.Directions.Direction[index + 1].Up) {
-                      rowAbility["links-top-2"] = true;
-                    }
-                    break;
-                  case "3":
-                    rowAbility.size = "triple";
-                    if (index < 2 && row.Directions.Direction[index + 1].Up) {
-                      rowAbility["links-top-2"] = true;
-                    }
-                    if (index < 2 && row.Directions.Direction[index + 2].Up) {
-                      rowAbility["links-top-3"] = true;
-                    }
-                    break;
-                  case "4":
-                    rowAbility.size = "full";
-                    if (index < 1 && row.Directions.Direction[index + 1].Up) {
-                      rowAbility["links-top-2"] = true;
-                    }
-                    if (index < 1 && row.Directions.Direction[index + 2].Up) {
-                      rowAbility["links-top-3"] = true;
-                    }
-                    if (index < 1 && row.Directions.Direction[index + 3].Up) {
-                      rowAbility["links-top-4"] = true;
-                    }
-                    break;
-                  default:
-                    rowAbility.size = "single";
-                    rowAbility.visible = false;
-                }
-
-                if (row.Directions.Direction[index].Right) {
-                  rowAbility["links-right"] = true;
-                }
-
-                const funcAddUpgradeDieModifier = (mod) => {
-                  if (Object.keys(CONFIG.temporary.skills).includes(mod.SkillKey)) {
-                    // only handling boosts initially
-                    if (mod.BoostCount || mod.SetbackCount || mod.AddSetbackCount || mod.ForceCount) {
-                      const skill = CONFIG.temporary.skills[mod.SkillKey];
-                      const modKey = Object.keys(rowAbility.attributes).length + 1;
-                      let modtype = "Skill Boost";
-                      let count = 0;
-                      if (mod.AddSetbackCount) {
-                        modtype = "Skill Setback";
-                        count = mod.AddSetbackCount;
-                      }
-                      if (mod.SetbackCount) {
-                        modtype = "Skill Remove Setback";
-                        count = mod.SetbackCount;
-                      }
-                      if (mod.ForceCount) {
-                        modtype = "Force Boost";
-                        count = true;
-                      }
-                      if (mod.BoostCount) {
-                        count = mod.BoostCount;
-                      }
-
-                      rowAbility.attributes[`attr${keyName}${modKey}`] = {
-                        mod: skill,
-                        modtype,
-                        value: count,
-                      };
-                    }
-                  }
-                };
-
-                if (rowAbilityData?.DieModifiers?.DieModifier) {
-                  if (Array.isArray(rowAbilityData.DieModifiers.DieModifier)) {
-                    rowAbilityData.DieModifiers.DieModifier.forEach((mod) => {
-                      funcAddUpgradeDieModifier(mod);
-                    });
-                  } else {
-                    if (Object.keys(CONFIG.temporary.skills).includes(rowAbilityData.DieModifiers.DieModifier.SkillKey)) {
-                      funcAddUpgradeDieModifier(rowAbilityData.DieModifiers.DieModifier);
-                    }
-                  }
-                }
-
-                const talentKey = `upgrade${(i - 1) * 4 + index}`;
-                power.data.upgrades[talentKey] = rowAbility;
-              });
-            } catch (err) {
-              CONFIG.logger.error(`Error importing record : `, err);
-            }
-          }
-
-          if (fp.ForcePower.AbilityRows.AbilityRow.length < 5) {
-            for (let i = fp.ForcePower.AbilityRows.AbilityRow.length; i < 5; i += 1) {
-              for (let index = 0; index < 4; index += 1) {
-                const talentKey = `upgrade${(i - 1) * 4 + index}`;
-
-                let rowAbility = { visible: false };
-
-                power.data.upgrades[talentKey] = rowAbility;
-              }
-            }
-          }
-
-          // does an image exist?
-          let imgPath = await ImportHelpers.getImageFilename(zip, "ForcePowers", "", power.flags.ffgimportid);
-          if (imgPath) {
-            power.img = await ImportHelpers.importImage(imgPath.name, zip, pack);
-          }
-
-          let compendiumItem;
-          await pack.getIndex();
-          let entry = pack.index.find((e) => e.name === power.name);
-
-          if (!entry) {
-            CONFIG.logger.debug(`Importing Force Power - Item`);
-            compendiumItem = new Item(power, { temporary: true });
-            this._importLogger(`New force power ${fp.ForcePower.Name} : ${JSON.stringify(compendiumItem)}`);
-            pack.importEntity(compendiumItem);
-          } else {
-            CONFIG.logger.debug(`Updating Force Power - Item`);
-            //let updateData = ImportHelpers.buildUpdateData(power);
-            let updateData = power;
-            updateData["_id"] = entry._id;
-            this._importLogger(`Updating force power ${fp.ForcePower.Name} : ${JSON.stringify(updateData)}`);
-            pack.updateEntity(updateData);
-          }
-          currentCount += 1;
-
-          $(".force .import-progress-bar")
-            .width(`${Math.trunc((currentCount / totalCount) * 100)}%`)
-            .html(`<span>${Math.trunc((currentCount / totalCount) * 100)}%</span>`);
-          this._importLogger(`End importing force power ${fp.ForcePower.Name}`);
-        } catch (err) {
-          CONFIG.logger.error(`Error importing record : `, err);
-        }
-      });
-    }
-    this._importLogger(`Completed Force Power Import`);
   }
 
   async _handleVehicles(zip) {
