+++ modules/importer/oggdude/oggdude.js
@@ -1,5 +1,6 @@
 import Armor from "./importers/armor.js";
 import Career from "./importers/careers.js";
+import ForcePowers from "./importers/forcepowers.js";
 import Gear from "./importers/gear.js";
 import ItemAttachments from "./importers/item-attachments.js";
 import ItemDescriptors from "./importers/item-descriptors.js";
@@ -14,6 +15,7 @@ export default class OggDude {
   static Import = {
     Armor: Armor.Import,
     Career: Career.Import,
+    ForcePowers: ForcePowers.Import,
     Gear: Gear.Import,
     ItemAttachments: ItemAttachments.Import,
     ItemDescriptors: ItemDescriptors.Import,
